+++ client/tarantar/space.h
@@ -48,7 +48,7 @@ int ts_space_fill(struct ts_spaces *s, struct ts_options *opts);
 
 struct ts_key*
 ts_space_keyalloc(struct ts_space *s, struct tnt_tuple *t, int fileid,
-                  int offset, int attach);
+                  uint64_t offset, int attach);
 
 void
 ts_space_keyfree(struct ts_space *s, struct ts_key *k);
